fix: overlay platform-specific settings when the value does not exist
